backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "anthropic[vertex]>=0.57.1",
    "arxiv>=2.2.0",
    "asyncio>=3.4.3",
    "fastapi>=0.115.14",
    "feedparser>=6.0.11",
    "google-cloud-translate>=3.21.1",
    "google-genai>=1.24.0",
    "google-generativeai>=0.8.5",
    "googletrans>=4.0.2",
    "openai>=1.95.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pylatex>=1.4.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "sqlmodel>=0.0.24",
    "sse-starlette>=2.4.1",
    "supabase>=2.16.0",
    "uvicorn>=0.35.0",
]
